requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
opencv-python-headless==4.9.0.80
ultralytics==8.2.0
torch==2.2.1
requests==2.31.0
python-multipart==0.0.9
gdown==5.1.0
langchain-core
watchdog==4.0.0
faster-whisper==0.10.0
ffmpeg-python==0.2.0
torchaudio==2.2.1
numpy==1.26.4
pandas==2.2.1
langchain-core==0.1.41
python-dotenv==1.0.1
ctranslate2
pymongo==4.6.3
#pip install --upgrade pymongo certifi
